Reworked Queued providers to use a shared compisition component to consolidate logic

+++ KHost.Common/Models/IModelWithPosition.cs
@@ -2,6 +2,6 @@
 {
     public interface IModelWithPosition
     {
-        public float Position { get; set; }
+        public int Position { get; set; }
     }
 }
